% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function1.R
\name{Disea_description}
\alias{Disea_description}
\title{Project_function1}
\usage{
Disea_description(
  data = ProjectData,
  Sc = "School",
  N = "Symptomatic",
  Ti = "Time_frame",
  TP = "True_positive",
  FP = "False_positive",
  TN = "True_negative",
  FN = "False_negative",
  P = "Population",
  PPV = "PositivePV",
  NPV = "NegatPV"
)
}
\arguments{
\item{data}{data used to apply the function}

\item{Sc}{the school's name where the outbreak is going on}

\item{N}{population symptomatic people detected through a syndromic algorithm}

\item{Ti}{time frame between the onset of the out break to the time when the diagnostic is made}

\item{TP}{true positive}

\item{FP}{False positive}

\item{TN}{true negative}

\item{FN}{false negative}

\item{P}{general population in the school}

\item{PPV}{positive predictive value}

\item{NPV}{negative predictive value}
}
\description{
Simultaneously calculates incidence prevalence sensitivity specificity positive and negative predictive values
}
\examples{
Disea_description()
}
\keyword{CD}
\keyword{box}
\keyword{count}
\keyword{data}
\keyword{load}
\keyword{plot}
\keyword{sacatter}
\keyword{viral}
\keyword{visualization}
