% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function3.R
\name{DrugsVisual}
\alias{DrugsVisual}
\title{Project_function3}
\usage{
DrugsVisual(data = drugLine)
}
\arguments{
\item{drugline}{data set encompassing variables medicine types viral load and CD4 count}
}
\description{
Data visualization smultaneously as scatter plots and boxplots
}
\examples{
MedSample()
}
\keyword{CD}
\keyword{box}
\keyword{count}
\keyword{data}
\keyword{load}
\keyword{plot}
\keyword{sacatter}
\keyword{viral}
\keyword{visualization}
