% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function2.R
\name{MedSample}
\alias{MedSample}
\title{Project_function2}
\usage{
MedSample(data = drugLine)
}
\arguments{
\item{drugline}{data set encompassing variables medicine types viral load and CD4 count}
}
\description{
Data visualization providing a scatter plots in on panel with two rows
}
\examples{
MedSample()
}
\keyword{CD}
\keyword{box}
\keyword{count}
\keyword{data}
\keyword{load}
\keyword{plot}
\keyword{sacatter}
\keyword{viral}
\keyword{visualization}
